Reverted rev. 495891 committed by Jacques Le Roux (based on a patch from Jonathon Wong, OFBIZ-620):
'Prepend feature idCodes with '-''

I'm sorry for this, I know that there has been a (rather long) discussion about this (very minor) patch, that has consumed a lot of time of some of the main committers (and now also mine!).
I don't like this code because it is too custom specific (why not '_' ?) and because it generates longer product ids (a bad thing if you have many different features) and because, if you want to prepend the '-' symbol to the idCodes you can actually add it to the idCodes.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/product@505229 13f79535-47bb-0310-9956-ffa450edef68