A patch from Mridul Pathak "Added support for uploading additional Product Views (images)" (https://issues.apache.org/jira/browse/OFBIZ-1930) - OFBIZ-1930
I fixed an, unrelated to patch, problem in EditProductContentContent.groovy. This is not the 1st time I cross this type of issue. I checked it does not appear on demo server and I guess it was not appearing with bsh  before. I'm not sure of its origin but it seems related to my configuration. I use, from my machine default, "Central European Summer Time" and demo server uses "Central Daylight Time". But I can't see why it affects my machine and not the demo server : there is no origin here (maybe GMT, but obviously it's not the case)
The lines I added are (+ import)
} else {
    fromDate = ObjectType.simpleTypeConvert(fromDate, "Timestamp", null, null, false)    
}
(I intentionnaly let the line without semi-colon as it's not mandatory in Groovy) 


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/product@689772 13f79535-47bb-0310-9956-ffa450edef68