Applied patch from jira issue: OFBIZ-3633 - Minimum order quantity. It will work as follows; We will set the special type price as 'MINIMUM_ORDER_PRICE' for a Product in ProductPrice entity. On the basis of it we will get the minimum order quantity of the product on the basis of this price and sale price. Will get the minimum order quantity for product by division. For example we have selling price of product P1 is $10.00 and its MINIMUM_ORDER_PRICE is $100.00 then minimum order quantity for the product will be --> 100/10 ==> 10.
To achieve the above ;
write a getMinimumOrderQuantity() method in ShoppingCart.java which takes the itemBasePrice and productId as in parameter, and call it where we add, update the cart items and change the quantity to minmumOrderQuantity if it is less then minimum.

Thanks Deepak & Rishi for the contribution and others for providing valuable comments / suggestions.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/product@936228 13f79535-47bb-0310-9956-ffa450edef68