Better Product French UI labels. I began by stumping upon the default page labels. The wording I certainly done some years ago without checking all changes in the UI was not really clear. So I wanted to remove the Id (référence in French) part but as we use really an Id there (and not a name or description or what not) I preferred to keep the Id concept. Then, while at it, I checked all the file to be consistent. I have not the courage to do it in all OFBiz :/
So the idea is to keep the Id words (référence, Réf. or sometimes [Réf.]) to make clear that in those places in OFBiz OOTB we use an Id. This because in some custom application we could prefer to use a name (for instance for lookup, or text field, display, etc.) and when you begin to mix the two things begin to be really less clear...

In one word, when you remove a "référence" word, please be careful... (sometimes it makes sense because nothing else could be there, for instance for UOM which are abbrev and may be considered name even if actually they are Id, but it's really subjective and I'd prefer to make the difference real clear: an Id is not a name and a name may be used too)

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/product@966525 13f79535-47bb-0310-9956-ffa450edef68