Fix "Be able to create a product from quick dropdown (Product Jump)" reported by BJ on dev ML at https://issues.apache.org/jira/browse/OFBIZ-4177 - OFBIZ-4177
BJ: The behavior was if you put in a productID that did not exist, and did a
jump to product, instead of a message No existing Products found, the
/catalog/control/EditProduct enter the ID.
then all you had to do was do an update to create the new product.
Is there a reason this functionality was removed?

JLR: It was leading to a bug for any other screen (but EditProductQuickAdmin and I just found that) than product creation, since there
was no related productId (try prices screen in stable demo for instance). It could be arranged for the product creation screen.

This patch gives this functionality back while keeping out the bug.
I have also removed redundant test on productId (since if there are no product I don't see the point to check for productId)



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/product@1069957 13f79535-47bb-0310-9956-ffa450edef68