Improvement - ProductContentWrapper didn't return ProductContent for virtual products (https://issues.apache.org/jira/browse/OFBIZ-4204).
Orig.Msg.:
Calling ProductContentWrapper.getProductContentAsText() for variant products:
When the method checks if "ProductContent" exists, it did not differ bedween variant and other products.
So if the ProductContent is empty for variant products, it did not check if the parentProduct has ProductContent of this type that could be returned.

This Patch improves the ProductContentWrapper to check variant products parent for ProductContent if thier content is empty.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/product@1078764 13f79535-47bb-0310-9956-ffa450edef68