A patch from Andreas König "EntityLabel support for Product Category Link Types needed." https://issues.apache.org/jira/browse/OFBIZ-4236
applications/product/data/ProductTypeData.xml defines an EnumerationType with enumTypeId="PCAT_LINK_TYPE".

The relevant elements next to this enumTypeId are used in the module "catalog / category / links" (see screenshot after applying the given patch).
Up to now we use the description given by the XML-Seeds above "hard-wired" in CategoryForms.xml, but it would be nice to have...

    * a way to add i18n support for the Product Category Link Types by using the ProductEntityLabels.xml and
    * support for a dynamic generation of the drop down list.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/product@1088017 13f79535-47bb-0310-9956-ffa450edef68