A patch from Andreas König "EditCategory.ftl is missing a "screenlet-body"-part showing the summery of an image-upload." https://issues.apache.org/jira/browse/OFBIZ-4244
After uploading a product-category-image, you receive the result of the upload-process in a screenlet.
Up to now, the screenlet has a div of the class "screenlet-title-bar", but no "screenlet-body".
To be consitent and to ensure a clear rendering, we should have the result-message in a div of the class "screenlet-body"

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/product@1100141 13f79535-47bb-0310-9956-ffa450edef68