A patch from Sebastian Leitner https://issues.apache.org/jira/browse/OFBIZ-4766 "PriceRule applies action to wrong priceType"
If you create a PriceRule and add an action of type "PRICE_POD" (Percent of default price) I read this as:
"take the default price and give an amount of x% to that"

These rules are applied within the service "calculateProductPrice" (specifically in PriceServices.calcPriceResultFromRules). It calculates the amount of discount using the default price (as specified by the action of the rule). But then it adds this discount to the listPrice instead of the defaultPrice.

An example:
listPrice = 100,
defaultPrice = 10,
rule says you give 10% of defaultPrice,
modifyAmount = -1 (this is the correct discount amount)
result would be 100-1 = 99.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/product@1308063 13f79535-47bb-0310-9956-ffa450edef68