A patch from Christian Carlow for "Shipment cancellation doesn't revert inventory accounting quantity total" https://issues.apache.org/jira/browse/OFBIZ-5314
When order items are issued to a shipment, the accounting_quantity_total
for the corresponding inventory records are updated but afterwards when
the shipment is cancelled, the accounting_quantity_total is not reverted
back to the original values.  Is this a bug or is there a reason the
quantity is not reverted?

The problem is that when I try to create another shipment and issue the
same quantity, I receive the following error:

The following error occurs when I try to issue the same previously
cancelled quantity to a newly created shipment:

Error:<br/> ERROR: Could not complete the Issue OrderItemShipGrpInvRes
to Shipment
[file:/dev/apache-ofbiz-11.04.02/applications/product/script/org/ofbiz/shipment/issuance/IssuanceServices.xml#issueOrderItemShipGrpInvResToShipment]
process [problem removing the orderItemShipGrpInvRes value: The current
transaction is marked for rollback, not beginning a new transaction and
aborting current operation; the rollbackOnly was caused by: Error in
simple-method [Create an accounting transactions for a sales shipment
issuance (D: INVENTORY_ACCOUNT, C: COGS_ACCOUNT)
[file:/home/stan/dev/apache-ofbiz-11.04.02/applications/accounting/script/org/ofbiz/accounting/ledger/GeneralLedgerServices.xml#createAcctgTransForSalesShipmentIssuance]]:
; {Could not find enough accounting inventory for product DOUGH;
remaining quantity: 1}] <br/> <br/>

The service is looking for inventory_item records with an
accounting_quantity_total > 0 but finds none because the quantity was
decremented to 0 from the cancelled shipment issuance.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/product@1531030 13f79535-47bb-0310-9956-ffa450edef68