A completed patch from Pierre Smits for "Using type=accounting-number in form in facility generates wrong output for inventory overview" https://issues.apache.org/jira/browse/OFBIZ-5335
When using locale DE or NL several quantities (ATP, QOH and others) get transformed wrong.
E.g. a ATP of 250 for a product is 250.00 in locale EN or FR, but for DE and NL the output is 250.000.000,00

jleroux: I completed Pierre's patch in 2 ways:
* used also BigDecimal for minimumStock in getProductInventoryAndFacilitySummary service and groovy script
* Fixed the same kind of issue in  CountFacilityInventoryByProduct.groovy

In by and large i18n contexts, this fixes number of Inventory views and reports, also in Asset Maint and Birt components where those 2 service/groovy scripts are used

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/product@1543766 13f79535-47bb-0310-9956-ffa450edef68