A patch from Youssef Khaye to finally fix "Incorrect quantityNotAvailable for OrderItemShipGrpInvRes when issuing items to shipments"   https://issues.apache.org/jira/browse/OFBIZ-5364
This patch fixes the problem from a functional point of view and consequently the JUnit test in question.

In fact, functionally speaking, The previous patch is not good in measure or it does not take into account that an OrderItem might use several lines of stock InventoryItem.
orderId     orderItemSeqId  orderItemShipGrpSeqId   InvetoryItemId
10000       000001              10000                               000001
10000       000001              10001                               000001

So by making a stock exit (ItemIssuance) we can afford to cancel the reservation lines for the line of stock in question but not all the lines of stock.
That's why I added the lines in the inventoryItemId research question and then cancel a reservation!

While at it I reverted r1653056 since testMultipleInventoryItemIssuance now passes

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/product@1656185 13f79535-47bb-0310-9956-ffa450edef68