Fix for "Print Pick Sheet PDF using orderId.pdf as attachement name" (OFBIZ-6398).
This was introduced by r1680671 during the improvement for (OFBIZ-4480). The behaviour is changed to name the attachment PickSheet<orderId>.pdf for single order pick sheets and preserve the original naming when used for multiple order pick sheets.

Thanks Deepak for reporting the issue and everyone else who took part in the discussion.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/product@1686536 13f79535-47bb-0310-9956-ffa450edef68