A patch from Florian Montalbano for "Stock Move: Unable to select From Stock location from List Facility Location popup" https://issues.apache.org/jira/browse/OFBIZ-7309 reported by Ankush Upadhyay
Functionality working as expected over ofbiz-13.07.03 but failing in trunk.
Steps to reproduce:
1. Login to facility component.
2. Select any facility.
2. Navigate to Stock Moves tab.
3. Click on From Location lookup from Quick Stock Move section against GZ-8544 product.
4. Try to select any location from List Facility Location popup.


We looked into the problem and found what seems to be the root of this bug.
The form with the not working lookup is described in the following ftl file : PickLiveStock.ftl .

For some reason, the 'Product' input and the 'To Location' input are created with the macro @htmlTemplate.lookupField (and are working well) but the 'From Location' field is created by a direct call to a javascript function (and is not working well).
That causes the lookup to open in another window and to lose the context. That's why the selection didn't work. Moreover, the theme was not rendered with Saphir.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/product@1748121 13f79535-47bb-0310-9956-ffa450edef68