Completes and closes "Use 'conditionFields' instead of 'andCondition' in Lookup screen" - https://issues.apache.org/jira/browse/OFBIZ-7655
In order to improve a little the readability, we can replace the use of direct call of groovy in Lookup screen for the autocompletion parameters. Instead, we can use the field 'conditionFields', which is a map key/value, to declare those parameters.

I found 2 remaining occurrences, this fixes them in one shoot, as I suggested in the main Jira (OFBIZ-7655)

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/product@1750543 13f79535-47bb-0310-9956-ffa450edef68