Fixed: product page missing form elements (OFBIZ-8489)
At r1746422, Deepak added a condition on EditProduct form. InventoryItemTypeId 
was enabled only for Product creation, else displayed as readonly, instead of 
editable dropdown

The only problem is that you can have the inventoryItemTypeId with a null value.
It's then considered by convention as not serialised. But then only the label 
shows. I fix it at by setting the inventoryItemTypeId as not serialised in this 
case. It was a bit tricky because you have to consider if the product is null 
(creation) and then not set product.inventoryItemTypeId. It would else defeat 
the creation case and give other contradictions.

Thanks: Wai for report

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/product@1766074 13f79535-47bb-0310-9956-ffa450edef68